Rename namespace.
